+++ packages/playwright-tests/wasm-split-harness/run.sh
