+++ packages/playwright-tests/wasm-split-harness/src/stars.js
