--- packages/wasm-split/wasm-split-harness/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "wasm-split-harness"
-version = "0.1.0"
-edition = "2021"
-publish = false
-authors = ["Jonathan Kelley"]
-
-[dependencies]
-dioxus = { workspace = true, features = ["web", "router", "wasm-split"] }
-anyhow = { workspace = true }
-async-compression = { workspace = true, features = ["futures-io", "gzip", "brotli"] }
-futures = { workspace = true }
-js-sys = { workspace = true }
-wasm-bindgen = { workspace = true }
-wasm-bindgen-futures = { workspace = true }
-web-sys = { workspace = true, features = ["Document", "Window", "HtmlElement", "Text", "DomRectReadOnly", "console"] }
-once_cell = { workspace = true }
-getrandom = { workspace = true, features = ["js"] }
-reqwest = { workspace = true, features = ["json"] }
-
